refactor: split server.go, manager.go, compiler.go (#371, #372, #373)
#371: Move session RPCs into server_sessions.go (1085 lines).
server.go drops from 9758 to 8698.

#372: Move HA/session sync methods into manager_ha.go (955 lines).
manager.go drops from 5387 to 4452.

#373: Move NAT compilation into compiler_nat.go (808 lines).
compiler.go drops from 6676 to 5878.

Closes #371 #372 #373
